requirements.txt
streamlit
pandas
plotly
numpy
streamlit-authenticator
duckdb
streamlit_option_menu
pillow
xlsxwriter